+++ R/aggregate.R
@@ -24,23 +24,15 @@ getColNameOfClass <- function(df, class_name){
 #'   Will auto perform mean aggregation it metric names match \code{mean_regex} argument
 #'   If agg_names is NULL will aggregate over all
 #'
-#' @importFrom dplyr %>%
-#' @importFrom dplyr select_
-#' @importFrom dplyr group_by_
-#' @importFrom dplyr summarise_each
-#' @importFrom dplyr funs
-#' @importFrom dplyr ungroup
-#' @importFrom dplyr left_join
-#' @importFrom dplyr mutate
-#' @importFrom dplyr bind_cols
+#' @importFrom magrittr %>%
 #' @export
 aggregateGAData <- function(ga_data, 
                             agg_names=NULL,
                             mean_regex="^avg|^percent|Rate$|^CPC$|^CTR$|^CPM$|^RPC$|^ROI$|^ROAS$|Per"){
   
-  stopifnot(inherits(ga_data, "data.frame"))
+  assertthat::assert_that(is.data.frame(ga_data))
   
-  if(!is.null(agg_names)) stopifnot(inherits(agg_names, "character"))
+  if(!is.null(agg_names)) assertthat::assert_that(inherits(agg_names, "character"))
   
   metrics <- getColNameOfClass(ga_data, "numeric")
   mean_metrics <- metrics[grepl(mean_regex, metrics)]
@@ -54,28 +46,29 @@ aggregateGAData <- function(ga_data,
   date_selects <- lapply(c(agg_names, date_col), as.symbol)
   
   dots <- lapply(agg_names, as.symbol)
-  
+
   ## metrics to take mean as per mean_regex
   meanAgg <- ga_data %>%
     dplyr::select_(.dots = mean_selects) %>%
     dplyr::group_by_(.dots=dots) %>%
-    dplyr::summarise_each(dplyr::funs(mean(., na.rm = TRUE))) %>% dplyr::ungroup()
+    dplyr::summarise_all(dplyr::funs(mean(., na.rm = TRUE))) %>% dplyr::ungroup()
   
   ## metrics to sum over
   sumAgg <- ga_data %>%
     dplyr::select_(.dots = sum_selects) %>%
     dplyr::group_by_(.dots=dots) %>%
-    dplyr::summarise_each(dplyr::funs(sum(., na.rm = TRUE))) %>% dplyr::ungroup()
+    dplyr::summarise_all(dplyr::funs(sum(., na.rm = TRUE))) %>% dplyr::ungroup()
   
   ## date dimensions take the first entry
   dateAgg <- ga_data %>%
     dplyr::select_(.dots = date_selects) %>%
     dplyr::group_by_(.dots=dots) %>%
-    dplyr::summarise_each(dplyr::funs(min(., na.rm = TRUE))) %>% dplyr::ungroup()
+    dplyr::summarise_all(dplyr::funs(min(., na.rm = TRUE))) %>% dplyr::ungroup()
   
   ## join up all the aggregations
   if(!is.null(agg_names)){
-    ga_agg <- dplyr::left_join(sumAgg, meanAgg) %>% dplyr::left_join(dateAgg)
+    ga_agg <- dplyr::left_join(sumAgg, meanAgg, by = agg_names) %>%
+      dplyr::left_join(dateAgg, by = agg_names)
   } else {
     sumAgg  <- if(ncol(sumAgg) == 0) NULL else sumAgg
     meanAgg <- if(ncol(meanAgg) == 0) NULL else meanAgg