+++ R/segments.R
@@ -30,7 +30,7 @@ ga_segment_list <- function(){
 #' @return a segmentFilter object. You can pass a list of these to the request.
 #' 
 #' @details 
-#' \code{segment_ga4} is the top heirarchy of segment creation, for which you will also need:
+#' \code{segment_ga4} is the top hierarchy of segment creation, for which you will also need:
 #' \itemize{
 #'  \item \link{segment_define} : AND combination of segmentFilters
 #'  \item \link{segment_vector_simple} or \link{segment_vector_sequence}
@@ -138,12 +138,12 @@ segment_ga4 <- function(name,
                         segment_id=NULL,
                         user_segment=NULL,
                         session_segment=NULL){
-  
-  testthat::expect_type(name, "character")
+
+  assertthat::assert_that(is.character(name))
   
   if(!is.null(segment_id)){
-    testthat::expect_type(segment_id, "character")
-    testthat::expect_length(segment_id, 1)
+    assertthat::assert_that(is.character(segment_id),
+                            length(segment_id) == 1)
   }
 
   
@@ -175,7 +175,7 @@ segment_ga4 <- function(name,
 #' Defines the segment to be a set of SegmentFilters 
 #'   which are combined together with a logical AND operation.
 #'   
-#' \code{segment_define} is in the heirarchy of segment creation, for which you will also need:
+#' \code{segment_define} is in the hierarchy of segment creation, for which you will also need:
 #' \itemize{
 #'  \item \link{segment_define} : AND combination of segmentFilters
 #'  \item \link{segment_vector_simple} or \link{segment_vector_sequence}
@@ -226,7 +226,7 @@ segment_define <- function(segment_filters,
 
 #' Make a simple segment vector
 #' 
-#' \code{segment_vector_simple} is in the heirarchy of segment creation, for which you will also need:
+#' \code{segment_vector_simple} is in the hierarchy of segment creation, for which you will also need:
 #' \itemize{
 #'  \item \link{segment_define} : AND combination of segmentFilters
 #'  \item \link{segment_vector_simple} or \link{segment_vector_sequence}
@@ -240,7 +240,7 @@ segment_define <- function(segment_filters,
 #' @export
 segment_vector_simple <- function(segment_elements){
   
-  testthat::expect_type(segment_elements, "list")
+  assertthat::assert_that(is.list(segment_elements))
   
   orFiltersList <- makeOrFilters(segment_elements)
   
@@ -261,7 +261,7 @@ makeOrFilters <- function(segment_element_list){
 #' Make sequenceSegment
 #' 
 #' 
-#' \code{segment_vector_sequence} is in the heirarchy of segment creation, for which you will also need:
+#' \code{segment_vector_sequence} is in the hierarchy of segment creation, for which you will also need:
 #' \itemize{
 #'  \item \link{segment_define} : AND combination of segmentFilters
 #'  \item \link{segment_vector_simple} or \link{segment_vector_sequence}
@@ -278,7 +278,7 @@ segment_vector_sequence <- function(segment_elements,
   
   stepMatchList <- lapply(segment_elements, function(x) attr(x, "matchType"))
   if(!is.null(stepMatchList)){
-    testthat::expect_equal(length(segment_elements), length(stepMatchList))
+    assertthat::assert_that(length(segment_elements) == length(stepMatchList))
   } else {
     stepMatchList <- as.list(rep("PRECEDES", length(segment_elements)))
   }
@@ -296,7 +296,7 @@ segment_vector_sequence <- function(segment_elements,
 
 #' Make a segment element
 #' 
-#' \code{segment_element} is the lowest heirarchy of segment creation, for which you will also need:
+#' \code{segment_element} is the lowest hierarchy of segment creation, for which you will also need:
 #' \itemize{
 #'  \item \link{segment_define} : AND combination of segmentFilters
 #'  \item \link{segment_vector_simple} or \link{segment_vector_sequence}
@@ -343,16 +343,16 @@ segment_element <- function(name,
                             comparisonValue=NULL,
                             matchType = c("PRECEDES", "IMMEDIATELY_PRECEDES")){
 
-  testthat::expect_type(name, "character")
+  assertthat::assert_that(is.character(name))
   operator <- match.arg(operator)
   type <- match.arg(type)
-  testthat::expect_type(not, "logical")
+  assertthat::assert_that(is.logical(not))
   expect_null_or_type(expressions, "character")
   expect_null_or_type(caseSensitive, "logical")  
-  expect_null_or_type(minComparisonValue, "double")  
-  expect_null_or_type(maxComparisonValue, "double")  
+  expect_null_or_type(minComparisonValue, "numeric")  
+  expect_null_or_type(maxComparisonValue, "numeric")  
   scope <- match.arg(scope)
-  expect_null_or_type(comparisonValue, "double")  
+  expect_null_or_type(comparisonValue, "numeric")  
   matchType <- match.arg(matchType)
   
   name <- sapply(name, checkPrefix, prefix = "ga")
@@ -370,7 +370,7 @@ segment_element <- function(name,
                               comparisonValue = as.character(comparisonValue),
                               maxComparisonValue = as.character(maxComparisonValue))
     
-    testthat::expect_s3_class(smf, "segmentMetFilter_ga4")
+    assertthat::assert_that(inherits(smf, "segmentMetFilter_ga4"))
     
     sfc <- segmentFilterClause(not = not, metricFilter = smf)
     
@@ -410,7 +410,7 @@ segment_element <- function(name,
                                  minComparisonValue = as.character(minComparisonValue),
                                  maxComparisonValue = as.character(maxComparisonValue))
     
-    testthat::expect_s3_class(sdf, "segmentDimFilter_ga4")
+    assertthat::assert_that(inherits(sdf, "segmentDimFilter_ga4"))
     sfc <- segmentFilterClause(not = not, dimensionFilter = sdf)
     
   }
@@ -454,8 +454,8 @@ segmentObj_ga4 <- function(dynamicSegment=NULL, segmentId=NULL){
 #'
 #' @keywords internal
 dynamicSegment <- function(name, userSegment, sessionSegment){
-  
-  testthat::expect_type(name, "character")
+
+  assertthat::assert_that(is.character(name))
   
   structure(
     list(
@@ -500,7 +500,7 @@ segmentDefinition <- function(segmentFilterList){
 #' @keywords internal
 segmentFilter <- function(not=FALSE, simpleSegment=NULL, sequenceSegment=NULL){
   
-  testthat::expect_type(not, "logical")
+  assertthat::assert_that(is.logical(not))
   
   if(!is.null(simpleSegment) && !is.null(sequenceSegment)){
     stop("Only one of either simpleSegment or sequenceSegment allowed")
@@ -644,7 +644,7 @@ segmentMetricFilter <- function(name,
 #' @keywords internal
 sequenceSegment <- function(segmentSequenceStepList, firstStepMatch=FALSE){
   
-  testthat::expect_type(firstStepMatch, "logical")
+  assertthat::assert_that(is.logical(firstStepMatch))
   
   structure(
     list(
@@ -658,7 +658,7 @@ sequenceSegment <- function(segmentSequenceStepList, firstStepMatch=FALSE){
 #' segmentSequenceStep
 #'
 #' @param orFiltersForSegmentList A list of \link{orFiltersForSegment}
-#' @param matchType Should it precede or immediatly precede the next step
+#' @param matchType Should it precede or immediately precede the next step
 #' 
 #' @return segmentSequenceStep object
 #'